Introduce discovery revisions to filter multiple revisioned configs
This PR allows pilot to watch for multiple revisioned
configurations at once, so that the configs can be propagated
to the connected proxies. This helps in propagating previous
revisioned configs to upgraded proxies.

The changes in the commit involves path of least resistance,
to carry forward the patch locally in the fork - till the time
a concrete solution is provided in upstream.

---------

Signed-off-by: Shubham Chauhan <shubham@tetrate.io>
(cherry picked from commit f48d3439c520b9a441a258c85eb2d42e40cc14e1)